+++ src/nodejs/tlid/tlid.js
@@ -19,6 +19,97 @@ function tlid__get() {
   return moment().format("YYMMDDHHMMSS");
 }
 
+
+var debugDt = false;
+/** Return a Date from a Tlid String
+ * @param {string} tlidStr
+ * */
+function tlid__toDate(str) {
+  // var dt = new Date(str); //@s 181223 Wont work as it require "-"
+  var tlidValue = tlid__xtr(str);
+  var r = "";
+
+  var y = "20" + getchk_sub(tlidValue, 0, 2);
+
+  var m = getchk_sub(tlidValue, 2, 4) - 1;
+  var d = getchk_sub(tlidValue, 4, 6);
+
+  r = `${y}/${m != "" ? m : 1}/${d != "" ? d : 1}`;
+
+  if (debugDt) console.log(`Year: ${y}`);
+  if (debugDt) console.log(`Month: ${m}`);
+  if (debugDt) console.log(`Day: ${d}`);
+  var h = "00";
+  var M = "00";
+  var s = "00";
+  var ms = "000";
+
+  try {
+
+    h = getchk_sub(tlidValue, 6, 8);
+    if (debugDt) console.log(`Hour: ${h}`);
+
+
+
+
+    M = getchk_sub(tlidValue, 8, 10);
+    if (debugDt) console.log(`Minutes: ${M}`);
+
+
+
+    s = getchk_sub(tlidValue, 10, 12);
+    if (debugDt) console.log(`Seconds: ${s}`);
+
+
+
+    ms = getchk_sub(tlidValue, 12, 13);
+    ms = getchk_sub(tlidValue, 12, 14);
+    ms = getchk_sub(tlidValue, 12, 15);
+    ms = getchk_sub(tlidValue, 12, 16);
+
+    if (debugDt) console.log(`MiliSeconds: ${ms}`);
+
+
+  } catch (error) {
+
+  }
+  if (debugDt) console.log("R: " + r);
+
+  var dt = new Date(r);
+
+  if (y != "")
+    if (m != "")
+      if (d != "")
+        if (h != "")
+          if (M != "")
+            if (s != "")
+              if (ms != "")
+                dt = new Date(Date.UTC(y, m, d, h, M, s, ms));
+              else
+                dt = new Date(Date.UTC(y, m, d, h, M, s));
+            else
+              dt = new Date(Date.UTC(y, m, d, h, M));
+          else
+            dt = new Date(Date.UTC(y, m, d, h));
+        else
+          dt = new Date(Date.UTC(y, m, d));
+      else
+        dt = new Date(Date.UTC(y, m));
+    else
+      dt = new Date(Date.UTC(y));
+
+  if (debugDt) console.log("TZ: " + dt.getTimezoneOffset());
+
+  return dt;
+}
+
+function getchk_sub(str, s, e) {
+  try {
+    return str.substring(s, e);
+  } catch (error) {
+    return "";
+  }
+}
 /**
  * JSON Description Quoted only
  */
@@ -108,8 +199,10 @@ function tlid__day() {
 
   if (tlid__has(str)) {
     r.tlid = tlid__xtr(str);
-    r.txt = str.replace(r.tlid, "")
-      .replace("@tlid ", "");
+    r.txt = str.replace(r.tlid + " ", "") // try to clear out the double space
+      .replace(r.tlid, "") // if ending the string, well we remove it
+      .replace("@tlid ", ""); //remove the decorator
+
   }
 
   return r;
@@ -133,8 +226,13 @@ function tlid__day() {
  */function
   tlid__clean(str) {
 
-  var numberPattern = /\d+/g;
-  return (str.replace(numberPattern, ""));
+
+  var tlidValue = tlid__xtr(str);
+  return str.replace(tlidValue + " ", "")
+    .replace(tlidValue, "")
+    .replace("@tlid ", "");
+
+
 }
 
 // tlid.js
@@ -233,6 +331,13 @@ try {
      */
     xtro: function (str) {
       return tlid__xtro(str);
+    },
+    /** Convert input tlid to dt
+     * 
+     * @param {*} str 
+     */
+    todate: function (str) {
+      return tlid__toDate(str);
     }
     ,
     /** Line has a tlid